pyproject.toml
[project]
name = "inkflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.100.0",
    "httpx>=0.28.1",
    "ollama>=0.6.2",
    "openai>=2.36.0",
    "pydantic>=2.13.4",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.20",
]

[tool.setuptools.packages.find]
include = ["inkflow*"]
